% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{caterpillar.dalmatian}
\alias{caterpillar.dalmatian}
\title{Caterpillar (dalmatian)}
\usage{
\method{caterpillar}{dalmatian}(
  object,
  family = NULL,
  nstart = start(object$coda),
  nend = end(object$coda),
  nthin = thin(object$coda),
  show = TRUE,
  return_plots = TRUE,
  ...
)
}
\arguments{
\item{object}{Object of class \code{dalmatian} created by \code{dalmatian()}.}

\item{family}{String defining selected family of variables (see help for \code{ggs()}).}

\item{nstart}{Start point for computing summary statistics (relative to true start of chain).}

\item{nend}{End point for computing summary statistics (relative to true start of chain).}

\item{nthin}{Thinning factor for computing summary statistics (relative to full chain and not previously thinned output).}

\item{show}{If TRUE then plots are displayed on the computer screen and the session is paused between each plot.}

\item{return_plots}{If TRUE then a named list of ggplot objects containing the plots will be returned as output.}

\item{...}{Ignored}
}
\value{
A list of \code{ggplot} objects that can be used to later reproduce the plots via \code{print}.
}
\description{
Construct caterpillar plots for key (or selected) parameters in a dalmatian object.
}
\references{
Bonner, S., Kim, H., Westneat, D., Mutzel, A.,
Wright, J., and Schofield, M.. (2021). \code{dalmatian}: A Package
for Fitting Double Hierarchical Linear Models in \code{R} via \code{JAGS} and
\code{nimble}. \emph{Journal of Statistical Software}, 100, 10, 1--25.
\doi{10.18637/jss.v100.i10}.
}
