% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{caterpillar}
\alias{caterpillar}
\title{Caterpillar (Generic)}
\usage{
caterpillar(object, ...)
}
\arguments{
\item{object}{Object to assess.}

\item{...}{Ignored}
}
\description{
Caterpillar (Generic)
}
\examples{

## Load output from previously run model
load(system.file("Pied_Flycatchers_1","pfresults.RData",package="dalmatian"))

## Generate caterpillar
pfcaterpillar <- caterpillar(pfresults,plot = FALSE)

}
