% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{gamma_data_1}
\alias{gamma_data_1}
\title{Simulated data to show how the gamma model may be fit with fixed and random effects on both the mean and dispersion.}
\format{
A data frame containing 1500 observations and 5 columns:
\describe{
  \item{ID}{The individual ID.}
  \item{Rep}{The replicate number.}
  \item{x1}{The value of the covariate for the mean.}
  \item{x2}{The value of the covariate for the dispersion.}
  \item{y}{The response.}
}
}
\usage{
gamma_data_1
}
\description{
Simulated data to show how the gamma model may be fit with fixed and random effects on both the mean and dispersion.
}
\keyword{datasets}
