% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{pfdata}
\alias{pfdata}
\title{Pied flycatcher feeding data}
\format{A data frame containing 5795 rows and 17 variables}
\usage{
pfdata
}
\description{
Dataset containing 5795 records of 60 pied flycatchers from 33 nest boxes feeding their nestlings during a brood manipulation experiment.
}
\keyword{datasets}
