% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict.dalmation}
\alias{predict.dalmation}
\title{Prediction Method for dalmation Fitted Objects}
\usage{
\method{predict}{dalmation}(object, df, method = "mean", ci = TRUE,
  level = 0.95)
}
\arguments{
\item{level}{}
}
\description{
Prediction Method for dalmation Fitted Objects
}
\examples{

}

