% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dalmatian.R
\name{print.dalmatian}
\alias{print.dalmatian}
\title{Printed Summary of a \code{dalmatian} Object}
\usage{
\method{print}{dalmatian}(x, summary = TRUE, convergence = TRUE, ...)
}
\arguments{
\item{x}{Object of class \code{dalmatian} created by \code{dalmatian()}.}

\item{summary}{If TRUE (default) compute posterior summary statistics via \code{summary.dalmatian()}.}

\item{convergence}{If TRUE (default) compute MCMC convergence diagnostics via \code{convergence()}.}

\item{...}{Ignored}
}
\value{
List of two elements containing posterior summary statstics and convergence diagnostics (if requested).
}
\description{
Prints summary information about a fitted model of class \code{dalmatian}.
}
\details{
This function produces a description of the model's structure and (by default) computes and prints the summary statistics
computed via \code{summary.dalmatian()} and the MCMC convergence diagnostics computed via \code{convergence.dalmatian()}.
Further control is available by calling these functions directly.
}
\examples{
\dontrun{
## Print summary of dalmatian objects
print(pfresults)
print(pfresults2)
}
}
\author{
Simon Bonner
}
