% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{print.dalmatian.summary}
\alias{print.dalmatian.summary}
\title{Print Summary (dalmatian)}
\usage{
\method{print}{dalmatian.summary}(x, digits = 2, ...)
}
\arguments{
\item{x}{Object of class \code{dalamtion.summary} created by \code{summary.dalmatian()}.}

\item{digits}{Number of digits to display after decimal.}

\item{...}{Ignored}
}
\description{
Print Summary (dalmatian)
}
