% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{print.dalmation.summary}
\alias{print.dalmation.summary}
\title{Print Summary (dalmation)}
\usage{
\method{print}{dalmation.summary}(x, digits = 2, ...)
}
\arguments{
\item{x}{Object of class \code{dalamtion.summary} created by \code{summary.dalmation()}.}

\item{digits}{Number of digits to display after decimal.}

\item{...}{Ignored}
}
\description{
Print Summary (dalmation)
}

