% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.dalmatian}
\alias{residuals.dalmatian}
\title{Residuals method for dalmatian fitted objects}
\usage{
\method{residuals}{dalmatian}(object, ...)
}
\arguments{
\item{object}{Object of class \code{dalmatian} created by \code{dalmatian()}.}

\item{...}{Ignored}
}
\value{
Data frame containing original data augmented with posterior
mean and lower and upper bounds of the 95% credible interval of the
residual for each observation.
}
\description{
Computes posterior summaries of the residuals for each observation.
Summary statistics include the posterior mean and the upper and
lower bounds of the 95% credible interval.
}
\examples{

## Load pied flycatcher data
data(pied_flycatchers_1)

## Create variables bounding the true load
pfdata$lower=ifelse(pfdata$load==0,log(.001),log(pfdata$load-.049))
pfdata$upper=log(pfdata$load+.05)

## Add 'log(IVI)' variable in pfdata
pfdata$'log(IVI)' <- log(pfdata$IVI)

## Load output from previously run model
load(system.file("Pied_Flycatchers_1","pfresults.RData",package="dalmatian"))

## Compute residuals
res.pfresults <- residual(object = pfresults)

}
\author{
Simon Bonner
}
