% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_jags_inits.R
\name{setJAGSInits}
\alias{setJAGSInits}
\title{Set initial values for \code{dalmatian}}
\usage{
setJAGSInits(mean.model, variance.model, fixed.mean = NULL,
  fixed.variance = NULL, y = NULL, random.mean = NULL, sd.mean = NULL,
  random.variance = NULL, sd.variance = NULL)
}
\arguments{
\item{mean.model}{Model list specifying the structure of the mean. (list)}

\item{variance.model}{Model list specifyint the structure of the variance. (list)}

\item{fixed.mean}{Initial values for the fixed effects of the mean. (numeric)}

\item{fixed.variance}{Initial values for the fixed effects of the variance. (numeric)}

\item{y}{Inital values for the true response. This should only be specified if the \code{rounding = TRUE} in the main call to dalmatian.}

\item{random.mean}{Initial values for the random effects of the mean. (numeric)}

\item{sd.mean}{Initial values for the standard deviation of the random effects of the mean. (numeric)}

\item{random.variance}{Initial values for the random effects of the variance. (numeric}

\item{sd.variance}{Initial values for the standard deviation of the random effects of the variance. (numeric)}
}
\value{
inits (list)
}
\description{
Set initial values for \code{dalmatian}
}
\details{
Allows the user to set initial values for \code{dalmatian}. Any values
not specified will by initialized by \code{JAGS}.
}
\author{
Simon Bonner
}
