% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{summary.dalmatian}
\alias{summary.dalmatian}
\title{Summary (dalmatian)}
\usage{
\method{summary}{dalmatian}(object, nstart = start(object$coda),
  nend = end(object$coda), nthin = thin(object$coda), ...)
}
\arguments{
\item{object}{Object of class \code{dalmatian} created by \code{dalmatian()}.}

\item{nstart}{Start point for computing summary statistics (relative to true start of chain).}

\item{nend}{End point for computing summary statistics (relative to true start of chain).}

\item{nthin}{Thinning factor for computing summary statsitics (relative to full chain and not previously thinned output).}

\item{...}{Ignored}
}
\value{
output (list)
}
\description{
Summary (dalmatian)
}
\examples{

## Load output from previously run model
load(system.file("Pied_Flycatchers_1","pfresults.RData",package="dalmatian"))

## Compute numerical summaries
summary(pfresults)
}
