% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{summary.dalmation}
\alias{summary.dalmation}
\title{Summary (dalmation)}
\usage{
\method{summary}{dalmation}(object, nstart = start(object$coda),
  nend = end(object$coda), nthin = thin(object$coda), ...)
}
\arguments{
\item{object}{Object of class \code{dalmation} created by \code{dalmation()}.}

\item{nstart}{Start point for computing summary statistics (relative to true start of chain).}

\item{nend}{End point for computing summary statistics (relative to true start of chain).}

\item{nthin}{Thinning factor for computing summary statsitics (relative to full chain and not previously thinned output).}

\item{...}{Ignored}
}
\value{
output (list)
}
\description{
Summary (dalmation)
}

