% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{terms.dalmatian}
\alias{terms.dalmatian}
\title{Terms function for \code{dalmatian} objects}
\usage{
\method{terms}{dalmatian}(object, ...)
}
\arguments{
\item{object}{Object of class \code{dalmatian} created by \code{dalmatian()}.}

\item{...}{Further object passed directly to \code{terms}. Recycled for each model component.}
}
\value{
List of with two lists named mean and dispersion each containing \code{terms} objects
corresponding to the fixed and random components of that model component (if present).
}
\description{
terms (dalmatian)
}
\details{
Constructs a list of terms objects for each component of the model specified in the
input object.
}
\examples{
\dontrun{
## Extract the terms objects corresponding to the pied-flycatcher model without random effects
terms(pfresults)

## Extract the terms objects corresponding to the pied-flycatcher model with random effects
terms(pfresults2)
}
}
\author{
Simon Bonner
}
