% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{traceplots.dalmation}
\alias{traceplots.dalmation}
\title{Traceplots (dalmation)}
\usage{
\method{traceplots}{dalmation}(object, family = NULL,
  nstart = start(object$coda), nend = end(object$coda),
  nthin = thin(object$coda), plot = TRUE, ...)
}
\arguments{
\item{object}{Object of class \code{dalmation} created by \code{dalmation()}.}

\item{family}{String defining selected family of variables (see help for \code{ggs()}).}

\item{nstart}{Start point for computing summary statistics (relative to true start of chain).}

\item{nend}{End point for computing summary statistics (relative to true start of chain).}

\item{nthin}{Thinning factor for computing summary statsitics (relative to full chain and not previously thinned output).}

\item{plot}{If TRUE then generate plots. Otherwise, a list of \code{ggplot} objects will be returned.}

\item{...}{Ignored}
}
\value{
A list of \code{ggplot} objects that can be used to later reproduce the plots via \code{print}.
}
\description{
Construct traceplots for key (or selected) parameters in a dalmation object.
}

