% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade_dalmation.R
\name{upgradedalmatian}
\alias{upgradedalmatian}
\title{Upgrade dalmatian object}
\usage{
upgradedalmatian(input)
}
\arguments{
\item{input}{Existing object to upgrade}
}
\value{
output (dalmatian)
}
\description{
Upgrade dalmatian object
}
\details{
Upgrades dalmatian object for use with new postprocessing functions.
}
\author{
Simon Bonner
}
