% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade_dalmation.R
\name{upgradeDalmation}
\alias{upgradeDalmation}
\title{Upgrade dalmation object}
\usage{
upgradeDalmation(input)
}
\arguments{
\item{input}{Existing object to upgrade}
}
\value{
output (dalmation)
}
\description{
Upgrade dalmation object
}
\details{
Upgrades dalmation object for use with new postprocessing functions.
}
\author{
Simon Bonner
}

