% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{weights_data_1}
\alias{weights_data_1}
\title{Simulated data for illustrating the use of weights}
\format{
An object of class \code{data.frame} with 100 rows and 3 columns.
}
\usage{
weights_data_1
}
\description{
Simulated data for illustrating the use of weights in the particular case when the responses are averages of observed with different denominators
}
\details{
@format A data frame with 100 rows and 3 columns:
 \describe{
   \item{n}{The number of observations.}
   \item{x}{The common predictor value.}
   \item{y}{The mean response value.}
 }
}
\keyword{datasets}
