% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{betabin_data_1}
\alias{betabin_data_1}
\title{Simulated data for illustrating the beta-binomial model}
\format{
A data frame containing 500 observations and 6 columns:
\describe{
  \item{ID}{The individual ID.}
  \item{Rep}{The replicate number}.
  \item{x1}{The value of the covariate for the mean.}
  \item{x2}{The value of the covariate for the dispersion.}
  \item{m}{The number of Bernoulli trials for each observation.}
  \item{y}{The number of successes.}
}
}
\usage{
betabin_data_1
}
\description{
Simulated data to show how the beta-binomial model may be fit with fixed and random effects on both the mean and dispersion.
}
\keyword{datasets}
