% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{caterpillar}
\alias{caterpillar}
\title{Caterpillar (Generic)}
\usage{
caterpillar(object, ...)
}
\arguments{
\item{object}{Object to assess.}

\item{...}{Ignored}
}
\value{
A list of \code{ggplot} objects that can be used to later reproduce the plots via \code{print}.
}
\description{
Caterpillar (Generic)
}
