% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{convergence}
\alias{convergence}
\title{Convergence Diagnostics (S3 Generic)}
\usage{
convergence(object, ...)
}
\arguments{
\item{object}{Object to asses.}

\item{...}{Ignored}
}
\description{
Generic function for computing convergence diagnostics.
}
\examples{

## Load output from previously run model
load(system.file("Pied_Flycatchers_1","pfresults.RData",package="dalmatian"))

## Compute convergence diagnostics
pfconvergence <- convergence(pfresults)

}
