% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{convergence}
\alias{convergence}
\title{Convergence Diagnostics (S3 Generic)}
\usage{
convergence(object, ...)
}
\arguments{
\item{object}{Object to asses.}

\item{...}{Ignored}
}
\value{
List containing Gelman-Rubin and Raftery convergence diagnostics and effective sample sizes for the selected parameters. This information is used to diagnose convergence of the MCMC sampling algorithms.
}
\description{
Generic function for computing convergence diagnostics.
}
