% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict.dalmatian}
\alias{predict.dalmatian}
\title{Prediction Method for dalmatian Fitted Objects}
\usage{
\method{predict}{dalmatian}(object, df, method = "mean", ci = TRUE,
  level = 0.95, ...)
}
\arguments{
\item{object}{Object of class \code{dalmatian} created by \code{dalmatian()}.}

\item{df}{data frame containing predictor values to predict response variables. (data.frame)}

\item{method}{Method to construct the fitted model. Either \code{"mean"} or \code{"mode"} (character)}

\item{ci}{returning credible intervals for predictions if TRUE (logical)}

\item{level}{level of credible intervals for predictions (numeric)}

\item{...}{Ignored}
}
\value{
predictions (list)
}
\description{
Prediction Method for dalmatian Fitted Objects
}
