% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{print.dalmatian.summary}
\alias{print.dalmatian.summary}
\title{Print Summary (dalmatian)}
\usage{
\method{print}{dalmatian.summary}(x, digits = 2, ...)
}
\arguments{
\item{x}{Object of class \code{dalamtion.summary} created by \code{summary.dalmatian()}.}

\item{digits}{Number of digits to display after decimal.}

\item{...}{Ignored}
}
\value{
No return value. This function prints the summary of a dalmatian object in a nicely formatted manner.
}
\description{
Print Summary (dalmatian)
}
\references{
Bonner, S., Kim, H., Westneat, D., Mutzel, A.,
Wright, J., and Schofield, M.. (2021). \code{dalmatian}: A Package
for Fitting Double Hierarchical Linear Models in \code{R} via \code{JAGS} and
\code{nimble}. \emph{Journal of Statistical Software}, 100, 10, 1--25.
\doi{10.18637/jss.v100.i10}.
}
