% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{ranef}
\alias{ranef}
\title{Random Effects (S3 Generic)}
\usage{
ranef(object, ...)
}
\arguments{
\item{object}{Input object}

\item{...}{Ignored}
}
\description{
Generic function for exporting summaries of random effects.
}
