% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{summary.dalmatian}
\alias{summary.dalmatian}
\title{Summary (dalmatian)}
\usage{
\method{summary}{dalmatian}(
  object,
  nstart = start(object$coda),
  nend = end(object$coda),
  nthin = thin(object$coda),
  ...
)
}
\arguments{
\item{object}{Object of class \code{dalmatian} created by \code{dalmatian()}.}

\item{nstart}{Start point for computing summary statistics (relative to true start of chain).}

\item{nend}{End point for computing summary statistics (relative to true start of chain).}

\item{nthin}{Thinning factor for computing summary statistics (relative to full chain and not previously thinned output).}

\item{...}{Ignored}
}
\value{
output (list)
}
\description{
Summary (dalmatian)
}
\references{
Bonner, S., Kim, H., Westneat, D., Mutzel, A.,
Wright, J., and Schofield, M.. (2021). \code{dalmatian}: A Package
for Fitting Double Hierarchical Linear Models in \code{R} via \code{JAGS} and
\code{nimble}. \emph{Journal of Statistical Software}, 100, 10, 1--25.
\doi{10.18637/jss.v100.i10}.
}
