% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{traceplots}
\alias{traceplots}
\title{Traceplots (Generic)}
\usage{
traceplots(object, ...)
}
\arguments{
\item{object}{Object to assess.}

\item{...}{Ignored}
}
\description{
Traceplots (Generic)
}
\examples{

## Load output from previously run model
load(system.file("Pied_Flycatchers_1","pfresults.RData",package="dalmatian"))

## Generate traceplots
pftraceplots <- traceplots(pfresults)

}
